requirement.txt
djangorestframework
markdown
django-filter
Pillow
django-cors-headers
django-hosts
